Bug fixes + Kubernetes 1.3+ auto launches dns
See [1], as of Kubernetes 1.3, DNS is a built-in service
launched automatically using the addon manager cluster add-on.

Fixes a missing character in kube-controller-manager config
which prevented the service from even launching. Essential
to cluster function.

Fixes an issue with the apiserver not being attached to the
cluster-internal hostnames and thus signing failed.

Has been tested, and confirmed kubernetes-dashboard deploys
and appears to be working.

[1] http://kubernetes.io/docs/admin/dns/