Updates README!
This is a start, I've updated it so that nothing is incorrect.
Unfortunately, it is not as detailed as the original version.

It's tempting to defer to the CoreOS + Kubernetes guide as it
is more thourough than the scope of this readme  and this
code is basically just a Terraform wrapper around those instructions.